Coordinator: swapNode() should keep retrying leader election until it succeeds (#781)
During `SwapNode()` operation for rebalancing we're triggering a new
leader election. If that leader election fails, the coordinator is not
handling the retries (unlike the case where the leader election is
triggered by a node failure).

The `SwapNode()` should actually block and wait until the leader
election finally succeeds, so that we don't move on to swap other
shards.

---------

Signed-off-by: Matteo Merli <mmerli@apache.org>