Handle deletion of swapped out node in a background periodic task (#786)
After a `SwapNode()` initiated leader election, let's not wait until the
removed node has deleted the shard since this is delaying the leader
election time.

Instead, move the logic out to a periodic task with retries.

Note: the node is immediately moved to a different category in the
status and marked in the `PendingDeleteShardNodes` list. This is because
it will not participate anymore in the newTermQuorum.

---------

Signed-off-by: Matteo Merli <mmerli@apache.org>