fix: data lost by change ensemble (#798)
### Motivation

We will update the memory metadata in the current implementation of
change ensemble(swap node), then trigger the leader election, which will
persist the new ensemble to metadata without any precondition. If there
is any failure to update the elected, the ensemble will be changed and
might be changed next time.

Keeping moving shards without follower-caught-up validation will cause a
data loss issue.


### Modification

1.  Introduce the status of the follower caught up in the election.
2. Move the caught-up validation into election ownership, and check that
the followers caught up after the new leader was elected. which can also
be cancelled by `election#stop`.
3. Check the `IsReadyForChangeEnsemble` before changing the ensemble,
and fail fast if the precondition is not satisfied.
4. Implement `copy-on-write` for `UpdateShardMetadata ` to avoid data
racing.